# doc/swagger.json
{"swagger":"2.0","info":{"description":"E-Auction Buyer REST Service API","version":"1.0","title":"E-Auction Buyer Service API","termsOfService":"Terms of Service","contact":{"name":"Mohamed Yusuff","url":"https://github.com/iSoftech/e-auction-buyer-service/","email":"me@contact.com"},"license":{"name":"License of API","url":"https://github.com/iSoftech/e-auction-buyer-service/blob/main/LICENSE"}},"host":"localhost:6060","basePath":"/","tags":[{"name":"E-Auction Buyer REST Controller","description":"Buyer Controller"},{"name":"Health Check REST Controller","description":"Health Controller"}],"paths":{"/buyer/place-bid":{"post":{"tags":["E-Auction Buyer REST Controller"],"summary":"[US_03] Adds a new Buyer and Bid","operationId":"placeBidUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"bidBuyerRequest","description":"bidBuyerRequest","required":true,"schema":{"$ref":"#/definitions/BidBuyerRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BidBuyerResponse"}},"201":{"description":"Created"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Bid not found"},"500":{"description":"Internal Server Error"}}}},"/buyer/show-bids":{"get":{"tags":["E-Auction Buyer REST Controller"],"summary":"Show all Bids","operationId":"showBidsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Bid"}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Bid not found"},"500":{"description":"Internal Server Error"}}}},"/buyer/show-bids/{product-id}":{"get":{"tags":["E-Auction Buyer REST Controller"],"summary":"[For US_04] Show all Bids for a given Product","operationId":"showBidsForProductUsingGET","produces":["*/*"],"parameters":[{"name":"product-id","in":"path","description":"product-id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BidResponse"}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Bid not found"},"500":{"description":"Internal Server Error"}}}},"/buyer/{buyer-email}/place-bid":{"post":{"tags":["E-Auction Buyer REST Controller"],"summary":"[US_03_Enhanced] Adds a new Bid for an existing Buyer","operationId":"placeBidByBuyerUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"bidRequest","description":"bidRequest","required":true,"schema":{"$ref":"#/definitions/BidRequest"}},{"name":"buyer-email","in":"path","description":"buyer-email","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Bid"}},"201":{"description":"Created"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Bid not found"},"500":{"description":"Internal Server Error"}}}},"/buyer/{buyer-email}/show-bids":{"get":{"tags":["E-Auction Buyer REST Controller"],"summary":"Show all Bids for an existing Buyer","operationId":"showBidsByBuyerIdUsingGET","produces":["*/*"],"parameters":[{"name":"buyer-email","in":"path","description":"buyer-email","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Bid"}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Bid not found"},"500":{"description":"Internal Server Error"}}}},"/buyer/{buyer-email}/update-bid":{"put":{"tags":["E-Auction Buyer REST Controller"],"summary":"[US_05_Enhanced] Updates the Bid Amount for the given Product","operationId":"updateBidUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"bidRequest","description":"bidRequest","required":true,"schema":{"$ref":"#/definitions/BidRequest"}},{"name":"buyer-email","in":"path","description":"buyer-email","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Bid"}},"201":{"description":"Created"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Bid not found"},"500":{"description":"Internal Server Error"}}}},"/health/node-status":{"get":{"tags":["Health Check REST Controller"],"summary":"Returns the service node status","description":"The end-point returns the node status of the service.","operationId":"nodeStatusUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Bid":{"type":"object","properties":{"bidAmount":{"type":"string"},"buyerId":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"productId":{"type":"integer","format":"int32"}},"title":"Bid"},"BidBuyerRequest":{"type":"object","properties":{"bidRequest":{"$ref":"#/definitions/BidRequest"},"buyerRequest":{"$ref":"#/definitions/BuyerRequest"}},"title":"BidBuyerRequest"},"BidBuyerResponse":{"type":"object","properties":{"bid":{"$ref":"#/definitions/Bid"},"buyer":{"$ref":"#/definitions/Buyer"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]}},"title":"BidBuyerResponse"},"BidRequest":{"type":"object","properties":{"bidAmount":{"type":"string"},"productId":{"type":"integer","format":"int32"}},"title":"BidRequest"},"BidResponse":{"type":"object","properties":{"bidAmount":{"type":"string"},"buyer":{"$ref":"#/definitions/Buyer"},"id":{"type":"integer","format":"int32"},"productId":{"type":"integer","format":"int32"}},"title":"BidResponse"},"Buyer":{"type":"object","properties":{"address":{"type":"string"},"city":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"id":{"type":"integer","format":"int32"},"lastName":{"type":"string"},"phone":{"type":"string"},"pin":{"type":"integer","format":"int32"},"state":{"type":"string"}},"title":"Buyer"},"BuyerRequest":{"type":"object","properties":{"address":{"type":"string"},"city":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"pin":{"type":"integer","format":"int32"},"state":{"type":"string"}},"title":"BuyerRequest"},"ResponseEntity":{"type":"object","properties":{"body":{"type":"object"},"statusCode":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"statusCodeValue":{"type":"integer","format":"int32"}},"title":"ResponseEntity"}}}